simple-microservice-rest-client: step1
