+++ get-started-microservices-on-kubernetes/simple-microservice-rest-client/simple-microservice-client-old/keycloak/myrealm.json
