+++ get-started-microservices-on-kubernetes/simple-microservice-rest-client/simple-microservice-client-old/src/main/java/org/wildfly/examples/GettingStartedApplication.java
