+++ get-started-microservices-on-kubernetes/simple-microservice-rest-client/simple-microservice-client-old/src/main/resources/META-INF/microprofile-config.properties
@@ -1,0 +1,4 @@
+simple-microservice-server/mp-rest/uri=http://127.0.0.1:8180/hello
+org.eclipse.microprofile.rest.client.propagateHeaders=Authorization
+#mp.jwt.verify.publickey.location=META-INF/public.pem
+#mp.jwt.verify.issuer=quickstart-jwt-issuer
