+++ get-started-microservices-on-kubernetes/simple-microservice-rest-client/simple-microservice-client-old/src/main/webapp/WEB-INF/oidc.json
