+++ get-started-microservices-on-kubernetes/simple-microservice-rest-client/simple-microservice-client-old/src/main/webapp/index.html
@@ -1,0 +1,61 @@
+<!--
+  ~ Copyright The WildFly Authors
+  ~ SPDX-License-Identifier: Apache-2.0
+  -->
+<!doctype html>
+<html class="no-js" lang="en">
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
+    <title>Welcome to WildFly Application Server</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    <link rel="stylesheet" href="normalize.min.css">
+    <link rel="stylesheet" href="wildfly.css">
+    <link rel="shortcut icon" href="favicon.ico" type="image/x-icon"/>
+    <script>
+        (() => {
+            "use script";
+            window.addEventListener("load", () => {
+                const button = document.querySelector("button[type='submit']");
+                button.addEventListener("click", (event) => {
+                    event.preventDefault();
+                    let name = document.querySelector('#nameInput');
+                    name = encodeURIComponent(name.value);
+                    const result = document.querySelector('#result');
+                    fetch(`/hello/${name}`)
+                        .then(response => {
+                            if (response.status === 200) {
+                                response.text().then(content => result.textContent = content);
+                            } else {
+                                result.textContent = `Unexpected response ${response.status}`;
+                            }
+                        });
+                });
+            });
+        })();
+    </script>
+</head>
+<body>
+<div class="wrapper">
+    <div class="content">
+        <div class="logo">
+            <img src="wildfly_logo.png" alt="Logo" title="Logo"/>
+        </div>
+        <h1>Getting Started</h1>
+        <div id="interactive">
+            <form name="nameForm">
+                <label for="nameInput">Name: </label>
+                <input type="text" id="nameInput">
+                <button type="submit">Say Hello</button>
+            </form>
+            <p id="result">
+            </p>
+        </div>
+    </div>
+</div>
+<footer>
+  <p>This page is served by your application. The source is in <code>src/main/webapp/index.html</code>.</p>
+</footer>
+</body>
+</html>
