+++ get-started-microservices-on-kubernetes/simple-microservice-rest-client/simple-microservice-client-old/src/main/webapp/wildfly.css
@@ -1,0 +1,79 @@
+/*
+  ~ Copyright The WildFly Authors
+  ~ SPDX-License-Identifier: Apache-2.0
+*/
+
+html {
+    color: #222;
+    font-family: "Open Sans", Geneva, Verdana, Arial, sans-serif;
+    font-size: 1em;
+    line-height: 1.4;
+}
+
+::-moz-selection {
+    background: #b3d4fc;
+    text-shadow: none;
+}
+
+::selection {
+    background: #b3d4fc;
+    text-shadow: none;
+}
+
+body {
+    background: url(bkg.gif) repeat-x;
+    color: #5B5B5B;
+    margin-top: 100px;
+    text-align: center;
+}
+
+h1 {
+    padding-top: 20px;
+    color: #7b1e1e;
+}
+
+a {
+    text-decoration: none;
+    color: #5e8a9a;
+}
+
+a:hover {
+    text-decoration: underline;
+    color: #8ec6d9;
+}
+
+img {
+    border: 0;
+}
+
+code {
+    background-color: rgba(27, 31, 35, 0.05);
+    border-radius: 3px;
+    font-family: Consolas, Menlo, Courier, monospace;
+    font-size: 90%;
+    padding: 0.2em;
+}
+
+.wrapper {
+    margin-left: auto;
+    margin-right: auto;
+    width: 50em;
+    text-align: left;
+}
+
+.content {
+    margin: 0;
+    padding: 0;
+}
+
+.logo {
+    margin-bottom: 10px;
+}
+
+footer {
+    position:fixed;
+    bottom:0;
+    left:0;
+    width:100%;
+    height: 90px;
+}
