+++ get-started-microservices-on-kubernetes/simple-microservice-rest-client/simple-microservice-client-old/src/test/java/org/wildfly/examples/TokenUtil.java
