+++ get-started-microservices-on-kubernetes/simple-microservice-rest-client/simple-microservice-client/pom.xml
@@ -3,11 +3,10 @@
   ~ Copyright The WildFly Authors
   ~ SPDX-License-Identifier: Apache-2.0
   -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <groupId>org.wildfly.examples</groupId>
-    <artifactId>simple-microservice</artifactId>
+    <artifactId>simple-microservice-client</artifactId>
     <version>1.0.0-SNAPSHOT</version>
 
     <packaging>war</packaging>
@@ -36,10 +35,6 @@
 
         <!-- maven-compiler-plugin -->
         <maven.compiler.release>11</maven.compiler.release>
-
-        <version.bom.microprofile>33.0.2.Final</version.bom.microprofile>
-
-        <version.com.nimbusds.jose-jwt>9.24.4.redhat-00001</version.com.nimbusds.jose-jwt>
     </properties>
 
     <!--
@@ -116,10 +111,19 @@
                 <scope>import</scope>
             </dependency>
 
+            <!--Define the JUnit5 bom. WildFly BOM still contains JUnit4, so we have to declare a version here -->
+            <dependency>
+                <groupId>org.junit</groupId>
+                <artifactId>junit-bom</artifactId>
+                <version>5.10.1</version>
+                <type>pom</type>
+                <scope>import</scope>
+            </dependency>
+
             <dependency>
                 <groupId>org.wildfly.bom</groupId>
                 <artifactId>wildfly-microprofile</artifactId>
-                <version>${version.bom.microprofile}</version>
+                <version>${version.wildfly.bom}</version>
                 <type>pom</type>
                 <scope>import</scope>
             </dependency>
@@ -151,49 +155,6 @@
             <groupId>org.eclipse.microprofile.config</groupId>
             <artifactId>microprofile-config-api</artifactId>
             <scope>provided</scope>
-        </dependency>
-
-        <!--<dependency>
-            <groupId>org.eclipse.microprofile.jwt</groupId>
-            <artifactId>microprofile-jwt-auth-api</artifactId>
-            <scope>provided</scope>
-        </dependency>-->
-
-        <!--<dependency>
-            <groupId>org.jboss.resteasy</groupId>
-            <artifactId>resteasy-tracing-api</artifactId>
-            <version>2.0.1.Final</version>
-            <scope>provided</scope>
-        </dependency>-->
-
-        <!-- OIDC Token -->
-        <dependency>
-            <groupId>org.wildfly.security</groupId>
-            <artifactId>wildfly-elytron-http-oidc</artifactId>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>jakarta.servlet</groupId>
-            <artifactId>jakarta.servlet-api</artifactId>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- TokenUtil -->
-        <dependency>
-            <groupId>com.nimbusds</groupId>
-            <artifactId>nimbus-jose-jwt</artifactId>
-            <version>${version.com.nimbusds.jose-jwt}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>jakarta.json</groupId>
-            <artifactId>jakarta.json-api</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.parsson</groupId>
-            <artifactId>parsson</artifactId>
-            <scope>test</scope>
         </dependency>
     </dependencies>
 
@@ -232,7 +193,7 @@
                             <location>org.wildfly:wildfly-galleon-pack:${version.wildfly.bom}</location>
                         </feature-pack>
                         <feature-pack>
-                            <location>org.wildfly.cloud:wildfly-cloud-galleon-pack:7.0.0.Final</location>
+                            <location>org.wildfly.cloud:wildfly-cloud-galleon-pack:7.0.2.Final</location>
                         </feature-pack>
                     </feature-packs>
                     <layers>
@@ -240,8 +201,6 @@
                         <layer>cloud-server</layer>
                         <layer>microprofile-config</layer>
                         <layer>microprofile-rest-client</layer>
-                        <layer>microprofile-jwt</layer>
-                        <layer>elytron-oidc-client</layer>
                     </layers>
                 </configuration>
                 <executions>
@@ -273,30 +232,4 @@
             </plugin>
         </plugins>
     </build>
-
-    <profiles>
-        <profile>
-            <id>bare-metal</id>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.wildfly.plugins</groupId>
-                        <artifactId>wildfly-maven-plugin</artifactId>
-                        <version>${version.wildfly.maven.plugin}</version>
-                        <configuration>
-                            <packaging-scripts>
-                                <packaging-script>
-                                    <commands>
-                                        <command>
-                                            /socket-binding-group=standard-sockets:write-attribute(name=port-offset,value=200)
-                                        </command>
-                                    </commands>
-                                </packaging-script>
-                            </packaging-scripts>
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-    </profiles>
 </project>
