+++ get-started-microservices-on-kubernetes/simple-microservice-rest-client/simple-microservice-client/src/main/resources/META-INF/microprofile-config.properties
@@ -1,4 +1,2 @@
-simple-microservice-server/mp-rest/uri=http://127.0.0.1:8180/hello
-org.eclipse.microprofile.rest.client.propagateHeaders=Authorization
-#mp.jwt.verify.publickey.location=META-INF/public.pem
-#mp.jwt.verify.issuer=quickstart-jwt-issuer
+simple-microservice-server/mp-rest/uri=${simple-microservice-server-uri:http://127.0.0.1:8080/hello}
+simple-microservice-server/mp-rest/connectTimeout=3000
