+++ get-started-microservices-on-kubernetes/simple-microservice-rest-client/simple-microservice-server-old/Dockerfile
@@ -1,0 +1,4 @@
+ARG runtime_image=quay.io/wildfly/wildfly-runtime:latest
+FROM ${runtime_image}
+COPY --chown=jboss:root target/server $JBOSS_HOME
+RUN chmod -R ug+rwX $JBOSS_HOME
