+++ get-started-microservices-on-kubernetes/simple-microservice-rest-client/simple-microservice-server-old/README.adoc
