+++ get-started-microservices-on-kubernetes/simple-microservice-rest-client/simple-microservice-server-old/kubernetes/deployment-my-jaxrs-app.yaml
@@ -1,0 +1,34 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: my-jaxrs-app-deployment
+  labels:
+    app: my-jaxrs-app
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: my-jaxrs-app
+  template:
+    metadata:
+      labels:
+        app: my-jaxrs-app
+    spec:
+      containers:
+        - name: my-jaxrs-app
+          image: quay.io/tborgato/my-jaxrs-app
+          ports:
+            - containerPort: 8080
+            - containerPort: 9990
+          livenessProbe:
+            httpGet:
+              path: /health/live
+              port: 9990
+          readinessProbe:
+            httpGet:
+              path: /health/ready
+              port: 9990
+          startupProbe:
+            httpGet:
+              path: /health/started
+              port: 9990
