+++ get-started-microservices-on-kubernetes/simple-microservice-rest-client/simple-microservice-server-old/src/main/java/org/wildfly/examples/GettingStartedApplication.java
@@ -6,7 +6,9 @@ package org.wildfly.examples;
 
 import jakarta.ws.rs.ApplicationPath;
 import jakarta.ws.rs.core.Application;
+import org.eclipse.microprofile.auth.LoginConfig;
 
+@LoginConfig(authMethod="MP-JWT", realmName="myrealm")
 @ApplicationPath("/hello")
 public class GettingStartedApplication extends Application {
