+++ get-started-microservices-on-kubernetes/simple-microservice-rest-client/simple-microservice-server-old/src/main/java/org/wildfly/examples/GettingStartedEndpoint.java
@@ -11,9 +11,13 @@ import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
+import org.eclipse.microprofile.jwt.JsonWebToken;
 
 @Path("/")
 public class GettingStartedEndpoint {
+
+    @Inject
+    JsonWebToken jwt;
 
     @Inject
     private GettingStartedService service;
@@ -22,7 +26,10 @@ public class GettingStartedEndpoint {
     @Path("/{name}")
     @Produces(MediaType.TEXT_PLAIN)
     public Response sayHello(final @PathParam("name") String name) {
-        String response = service.hello(name);
+        String response =
+                service.hello(name)
+                        + " Subject:" + (jwt!=null? jwt.getSubject():null)
+                        + " Issuer: " + jwt.getIssuer();
 
         return Response.ok(response).build();
     }
