+++ get-started-microservices-on-kubernetes/simple-microservice-rest-client/simple-microservice-server-old/src/main/java/org/wildfly/examples/GettingStartedService.java
@@ -10,6 +10,6 @@ import jakarta.enterprise.context.ApplicationScoped;
 public class GettingStartedService {
 
     public String hello(String name) {
-        return String.format("Hello '%s'.", name);
+        return String.format("Helloooooo '%s'.", name);
     }
 }
