+++ get-started-microservices-on-kubernetes/simple-microservice-rest-client/simple-microservice-server/pom.xml
@@ -6,7 +6,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <groupId>org.wildfly.examples</groupId>
-    <artifactId>simple-microservice</artifactId>
+    <artifactId>simple-microservice-server</artifactId>
     <version>1.0.0-SNAPSHOT</version>
 
     <packaging>war</packaging>
@@ -35,8 +35,6 @@
 
         <!-- maven-compiler-plugin -->
         <maven.compiler.release>11</maven.compiler.release>
-
-        <version.bom.microprofile>33.0.2.Final</version.bom.microprofile>
     </properties>
 
     <!--
@@ -113,10 +111,11 @@
                 <scope>import</scope>
             </dependency>
 
+            <!--Define the JUnit5 bom. WildFly BOM still contains JUnit4, so we have to declare a version here -->
             <dependency>
-                <groupId>org.wildfly.bom</groupId>
-                <artifactId>wildfly-microprofile</artifactId>
-                <version>${version.bom.microprofile}</version>
+                <groupId>org.junit</groupId>
+                <artifactId>junit-bom</artifactId>
+                <version>5.10.1</version>
                 <type>pom</type>
                 <scope>import</scope>
             </dependency>
@@ -136,16 +135,6 @@
         <dependency>
             <groupId>jakarta.ws.rs</groupId>
             <artifactId>jakarta.ws.rs-api</artifactId>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.microprofile.config</groupId>
-            <artifactId>microprofile-config-api</artifactId>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.microprofile.jwt</groupId>
-            <artifactId>microprofile-jwt-auth-api</artifactId>
             <scope>provided</scope>
         </dependency>
     </dependencies>
@@ -185,14 +174,12 @@
                             <location>org.wildfly:wildfly-galleon-pack:${version.wildfly.bom}</location>
                         </feature-pack>
                         <feature-pack>
-                            <location>org.wildfly.cloud:wildfly-cloud-galleon-pack:7.0.0.Final</location>
+                            <location>org.wildfly.cloud:wildfly-cloud-galleon-pack:7.0.2.Final</location>
                         </feature-pack>
                     </feature-packs>
                     <layers>
                         <!-- layers may be used to customize the server to provision-->
                         <layer>cloud-server</layer>
-                        <layer>microprofile-config</layer>
-                        <layer>microprofile-jwt</layer>
                     </layers>
                 </configuration>
                 <executions>
@@ -224,30 +211,4 @@
             </plugin>
         </plugins>
     </build>
-
-    <profiles>
-        <profile>
-            <id>bare-metal</id>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.wildfly.plugins</groupId>
-                        <artifactId>wildfly-maven-plugin</artifactId>
-                        <version>${version.wildfly.maven.plugin}</version>
-                        <configuration>
-                            <packaging-scripts>
-                                <packaging-script>
-                                    <commands>
-                                        <command>
-                                            /socket-binding-group=standard-sockets:write-attribute(name=port-offset,value=100)
-                                        </command>
-                                    </commands>
-                                </packaging-script>
-                            </packaging-scripts>
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-    </profiles>
 </project>
